Mundane gems and enemies must handle their own description updates
Remove MouseDownScript, replaced with a script to debug click and select on any object
