+++ Gem Tower Defense/Assets/Prefabs/ChippedGems_0.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 3790817751405389727}
   - component: {fileID: 4823269728508851671}
   - component: {fileID: 1035229177024982712}
+  - component: {fileID: 114738908729303198}
   m_Layer: 0
   m_Name: ChippedGems_0
   m_TagString: Untagged
@@ -106,3 +107,17 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 1, y: 1}
   m_EdgeRadius: 0
+--- !u!114 &114738908729303198
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5112011046384422838}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 09d5bf152e5db4944bdabfbf7e3e95f7, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  type: 
+  description: {fileID: 0}
