+++ Gem Tower Defense/Assets/Prefabs/ChippedGems_3.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 3947950059783357296}
   - component: {fileID: 7279220873293690666}
   - component: {fileID: 4866811663903567214}
+  - component: {fileID: 114917054499922766}
   m_Layer: 0
   m_Name: ChippedGems_3
   m_TagString: Untagged
@@ -106,3 +107,17 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 1, y: 1}
   m_EdgeRadius: 0
+--- !u!114 &114917054499922766
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1133853169638611904}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 09d5bf152e5db4944bdabfbf7e3e95f7, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  type: 
+  description: {fileID: 0}
