+++ Gem Tower Defense/Assets/Prefabs/ChippedGems_5.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 6092680345708975539}
   - component: {fileID: 6816876640335281971}
   - component: {fileID: 6377432611463177925}
+  - component: {fileID: 114075095701408926}
   m_Layer: 0
   m_Name: ChippedGems_5
   m_TagString: Untagged
@@ -106,3 +107,17 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 1, y: 1}
   m_EdgeRadius: 0
+--- !u!114 &114075095701408926
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3540688706979768033}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 09d5bf152e5db4944bdabfbf7e3e95f7, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  type: 
+  description: {fileID: 0}
