+++ Gem Tower Defense/Assets/Prefabs/base_6.prefab
@@ -69,7 +69,7 @@ SpriteRenderer:
   m_AutoUVMaxAngle: 89
   m_LightmapParameters: {fileID: 0}
   m_SortingLayerID: -718715799
-  m_SortingLayer: 1
+  m_SortingLayer: 2
   m_SortingOrder: 0
   m_Sprite: {fileID: 21300012, guid: a886ef5111bf6654e8478c1493a815b0, type: 3}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
@@ -120,9 +120,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 8cfa604abca155b48a8329c48c8ac2bc, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  type: 
+  description: {fileID: 0}
   hp: 0
   speed: 0
-  description: {fileID: 0}
 --- !u!114 &114022792774016168
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -130,7 +131,7 @@ MonoBehaviour:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 4285739087781507221}
-  m_Enabled: 1
+  m_Enabled: 0
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 70ebf2b649c1d9e4ebcc9e5ce89d03a1, type: 3}
   m_Name:
