+++ Gem Tower Defense/Assets/Scenes/MainLevel.unity
@@ -149,6 +149,80 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 70ebf2b649c1d9e4ebcc9e5ce89d03a1, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!1001 &548179675
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 114627259305456414, guid: ce3de01d00c9a554c81cbbf4f2ab1799,
+        type: 3}
+      propertyPath: m_Enabled
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5999738374615160529, guid: ce3de01d00c9a554c81cbbf4f2ab1799,
+        type: 3}
+      propertyPath: m_Name
+      value: base_3
+      objectReference: {fileID: 0}
+    - target: {fileID: 7149994813414960457, guid: ce3de01d00c9a554c81cbbf4f2ab1799,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: -2.19
+      objectReference: {fileID: 0}
+    - target: {fileID: 7149994813414960457, guid: ce3de01d00c9a554c81cbbf4f2ab1799,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 9.39
+      objectReference: {fileID: 0}
+    - target: {fileID: 7149994813414960457, guid: ce3de01d00c9a554c81cbbf4f2ab1799,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7149994813414960457, guid: ce3de01d00c9a554c81cbbf4f2ab1799,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7149994813414960457, guid: ce3de01d00c9a554c81cbbf4f2ab1799,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7149994813414960457, guid: ce3de01d00c9a554c81cbbf4f2ab1799,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7149994813414960457, guid: ce3de01d00c9a554c81cbbf4f2ab1799,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 7149994813414960457, guid: ce3de01d00c9a554c81cbbf4f2ab1799,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 7149994813414960457, guid: ce3de01d00c9a554c81cbbf4f2ab1799,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7149994813414960457, guid: ce3de01d00c9a554c81cbbf4f2ab1799,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7149994813414960457, guid: ce3de01d00c9a554c81cbbf4f2ab1799,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: ce3de01d00c9a554c81cbbf4f2ab1799, type: 3}
 --- !u!1 &695288717
 GameObject:
   m_ObjectHideFlags: 0
@@ -481,7 +555,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 4
+  m_RootOrder: 6
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1250178626
 GameObject:
@@ -737,7 +811,7 @@ PrefabInstance:
     - target: {fileID: 7625164376214479814, guid: 02ae0f95d338f974187de7cc2cbe9ec5,
         type: 3}
       propertyPath: m_RootOrder
-      value: 2
+      value: 4
       objectReference: {fileID: 0}
     - target: {fileID: 7625164376214479814, guid: 02ae0f95d338f974187de7cc2cbe9ec5,
         type: 3}
@@ -852,7 +926,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 3
+  m_RootOrder: 5
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1001 &1477769820
 PrefabInstance:
@@ -904,7 +978,7 @@ PrefabInstance:
     - target: {fileID: 840557710161529978, guid: 226e50ece12675147995afe2fe05d6a9,
         type: 3}
       propertyPath: m_RootOrder
-      value: 1
+      value: 3
       objectReference: {fileID: 0}
     - target: {fileID: 840557710161529978, guid: 226e50ece12675147995afe2fe05d6a9,
         type: 3}
@@ -934,10 +1008,79 @@ PrefabInstance:
     - target: {fileID: 4968126354426001675, guid: 226e50ece12675147995afe2fe05d6a9,
         type: 3}
       propertyPath: m_SortingLayer
-      value: 1
+      value: 2
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 226e50ece12675147995afe2fe05d6a9, type: 3}
+--- !u!1001 &1546523220
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 5112011046384422838, guid: 7c862c548cb5a914da2b2e4e3c60b4c7,
+        type: 3}
+      propertyPath: m_Name
+      value: ChippedGems_0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3790817751405389727, guid: 7c862c548cb5a914da2b2e4e3c60b4c7,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: -2.1497846
+      objectReference: {fileID: 0}
+    - target: {fileID: 3790817751405389727, guid: 7c862c548cb5a914da2b2e4e3c60b4c7,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 8.261311
+      objectReference: {fileID: 0}
+    - target: {fileID: 3790817751405389727, guid: 7c862c548cb5a914da2b2e4e3c60b4c7,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3790817751405389727, guid: 7c862c548cb5a914da2b2e4e3c60b4c7,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3790817751405389727, guid: 7c862c548cb5a914da2b2e4e3c60b4c7,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3790817751405389727, guid: 7c862c548cb5a914da2b2e4e3c60b4c7,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3790817751405389727, guid: 7c862c548cb5a914da2b2e4e3c60b4c7,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3790817751405389727, guid: 7c862c548cb5a914da2b2e4e3c60b4c7,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 2
+      objectReference: {fileID: 0}
+    - target: {fileID: 3790817751405389727, guid: 7c862c548cb5a914da2b2e4e3c60b4c7,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3790817751405389727, guid: 7c862c548cb5a914da2b2e4e3c60b4c7,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3790817751405389727, guid: 7c862c548cb5a914da2b2e4e3c60b4c7,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 7c862c548cb5a914da2b2e4e3c60b4c7, type: 3}
 --- !u!1 &1793568914
 GameObject:
   m_ObjectHideFlags: 0
@@ -1033,7 +1176,7 @@ RectTransform:
   - {fileID: 695288718}
   - {fileID: 1061400877}
   m_Father: {fileID: 0}
-  m_RootOrder: 5
+  m_RootOrder: 7
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
